Remove extra code
